// src/modules/notification/email/constants/email.enum.ts
export enum EmailType {
  // Welcome emails for different user types
  WELCOME_ADMIN = 'WELCOME_ADMIN',
  WELCOME_USER = 'WELCOME_USER',

  // Authentication related emails
  PASSWORD_RESET = 'PASSWORD_RESET',
  EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
  EMAIL_VERIFICATION_SUCCESS = 'EMAIL_VERIFICATION_SUCCESS',
  LOGIN_REMINDER = 'LOGIN_REMINDER',
  ACCOUNT_ACTIVATION = 'ACCOUNT_ACTIVATION',

  ACCOUNT_DEACTIVATION = 'ACCOUNT_DEACTIVATION',

  // Newsletter and marketing emails
  NEWSLETTER_SUBSCRIPTION = 'NEWSLETTER_SUBSCRIPTION',
  WAITLIST_SUBSCRIPTION = 'WAITLIST_SUBSCRIPTION',
  SUBSCRIPTION_RENEWAL = 'SUBSCRIPTION_RENEWAL',

  // Payments
  PAYMENT_STATUS_UPDATE = 'PAYMENT_STATUS_UPDATE',

  // Contact form
  CONTACT_FORM_SUBMISSION = 'CONTACT_FORM_SUBMISSION',

  // Marketplace
  PRODUCT_AVAILABILITY_NOTIFICATION = 'PRODUCT_AVAILABILITY_NOTIFICATION',

  // Tracking
  TRACKING_STATUS_UPDATE = 'TRACKING_STATUS_UPDATE',
  ORDER_DISPATCHED = 'ORDER_DISPATCHED',
  ORDER_IN_TRANSIT = 'ORDER_IN_TRANSIT',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_EXCEPTION = 'ORDER_EXCEPTION',
}